move priorities definitions to modular implementation
