+++ generators/base/generator-base.cjs
@@ -71,7 +71,7 @@ const { CONSUL, EUREKA } = require('../../jdl/jhipster/service-discovery-types')
 const { GATLING, CUCUMBER, PROTRACTOR, CYPRESS } = require('../../jdl/jhipster/test-framework-types');
 const { GATEWAY, MICROSERVICE, MONOLITH } = require('../../jdl/jhipster/application-types');
 const { ELASTICSEARCH } = require('../../jdl/jhipster/search-engine-types');
-const { CUSTOM_PRIORITIES } = require('../../lib/constants/priorities.cjs');
+const { CUSTOM_PRIORITIES } = require('./priorities.cjs');
 const cacheTypes = require('../../jdl/jhipster/cache-types');
 const serviceDiscoveryTypes = require('../../jdl/jhipster/service-discovery-types');
 const searchEngineTypes = require('../../jdl/jhipster/search-engine-types');
