+++ generators/base/priorities.mjs
@@ -16,31 +16,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import priorities from './priorities.cjs';
 
-export {
-  PRIORITY_PREFIX,
-  PRIORITY_NAMES,
-  BASE_PRIORITY_NAMES,
-  ENTITY_PRIORITY_NAMES,
-  INITIALIZING_PRIORITY,
-  PROMPTING_PRIORITY,
-  CONFIGURING_PRIORITY,
-  COMPOSING_PRIORITY,
-  LOADING_PRIORITY,
-  PREPARING_PRIORITY,
-  CONFIGURING_EACH_ENTITY_PRIORITY,
-  LOADING_ENTITIES_PRIORITY,
-  PREPARING_EACH_ENTITY_PRIORITY,
-  PREPARING_EACH_ENTITY_FIELD_PRIORITY,
-  PREPARING_EACH_ENTITY_RELATIONSHIP_PRIORITY,
-  POST_PREPARING_EACH_ENTITY_PRIORITY,
-  DEFAULT_PRIORITY,
-  WRITING_PRIORITY,
-  WRITING_ENTITIES_PRIORITY,
-  POST_WRITING_PRIORITY,
-  POST_WRITING_ENTITIES_PRIORITY,
-  PRE_CONFLICTS_PRIORITY,
-  INSTALL_PRIORITY,
-  POST_INSTALL_PRIORITY,
-  END_PRIORITY,
-} from './priorities.cjs';
+const { PRIORITY_PREFIX, QUEUE_PREFIX, PRIORITY_NAMES, PRIORITY_NAMES_LIST, QUEUES, CUSTOM_PRIORITIES } = priorities;
+
+export { PRIORITY_PREFIX, QUEUE_PREFIX, PRIORITY_NAMES, PRIORITY_NAMES_LIST, QUEUES, CUSTOM_PRIORITIES };
