+++ generators/generate-blueprint/constants.mjs
@@ -19,9 +19,9 @@
 import chalk from 'chalk';
 
 import GENERATOR_LIST from '../generator-list.cjs';
-import { BASE_PRIORITY_NAMES, ENTITY_PRIORITY_NAMES } from '../../lib/constants/priorities.mjs';
+import { PRIORITY_NAMES_LIST } from '../base-application/priorities.mjs';
 
-const prioritiesForSub = subGenerator => (subGenerator.startsWith('entit') ? ENTITY_PRIORITY_NAMES : BASE_PRIORITY_NAMES);
+const prioritiesForSub = subGenerator => PRIORITY_NAMES_LIST;
 
 export const GENERATE_SNAPSHOTS = 'generateSnapshots';
 export const LINK_JHIPSTER_DEPENDENCY = 'linkJhipsterDependency';
