+++ test/support/index.cjs
@@ -5,16 +5,18 @@ const { existsSync } = require('fs');
 
 const { GENERATOR_JHIPSTER } = require('../../generators/generator-constants.cjs');
 const { skipPrettierHelpers: helpers } = require('../utils/utils.cjs');
+const { PRIORITY_NAMES, ENTITY_PRIORITY_NAMES, PRIORITY_NAMES_LIST } = require('../../generators/base-application/priorities.cjs');
+
 const {
-  BASE_PRIORITY_NAMES,
-  BASE_ENTITY_PRIORITY_NAMES,
-  compat: {
-    CONFIGURING_EACH_ENTITY_PRIORITY,
-    PREPARING_EACH_ENTITY_PRIORITY,
-    PREPARING_EACH_ENTITY_FIELD_PRIORITY,
-    PREPARING_EACH_ENTITY_RELATIONSHIP_PRIORITY,
-  },
-} = require('../../lib/constants/priorities.cjs');
+  CONFIGURING_EACH_ENTITY,
+  LOADING_ENTITIES,
+  PREPARING_EACH_ENTITY,
+  PREPARING_EACH_ENTITY_FIELD,
+  PREPARING_EACH_ENTITY_RELATIONSHIP,
+  POST_PREPARING_EACH_ENTITY,
+  WRITING_ENTITIES,
+  POST_WRITING_ENTITIES,
+} = PRIORITY_NAMES;
 
 const testOptions = data => {
   const { generatorPath, customOptions, contextBuilder = () => helpers.create(generatorPath) } = data;
@@ -183,7 +185,7 @@ const testBlueprintSupport = (generatorName, options = {}) => {
     const prioritiesSpy = sinon.spy();
     const prioritiesTasks = [];
     let prioritiesCount = 0;
-    [...BASE_PRIORITY_NAMES, ...(entity ? BASE_ENTITY_PRIORITY_NAMES : [])].forEach(priority => {
+    PRIORITY_NAMES_LIST.forEach(priority => {
       let callback;
       if (Object.getOwnPropertyDescriptor(Object.getPrototypeOf(generator), `${taskPrefix}${priority}`)) {
         prioritiesCount++;
@@ -296,7 +298,7 @@ const testBlueprintSupport = (generatorName, options = {}) => {
     it('should call every priority', () => {
       expect(spy.prioritiesSpy.callCount).toBe(spy.prioritiesCount);
     });
-    BASE_PRIORITY_NAMES.forEach(priority => {
+    PRIORITY_NAMES_LIST.filter(priority => !Object.values(ENTITY_PRIORITY_NAMES).includes(priority)).forEach(priority => {
       it(`should call ${priority} tasks if implemented`, function () {
         if (!spy.prioritiesTasks[priority]) {
           this.skip();
@@ -306,17 +308,37 @@ const testBlueprintSupport = (generatorName, options = {}) => {
       });
     });
     if (entity) {
-      it(`should call ${CONFIGURING_EACH_ENTITY_PRIORITY} tasks twice`, function () {
-        expect(spy.prioritiesTasks[CONFIGURING_EACH_ENTITY_PRIORITY].callCount).toBe(2);
+      [LOADING_ENTITIES, WRITING_ENTITIES, POST_WRITING_ENTITIES].forEach(priority => {
+        it(`should call ${priority} tasks once`, function () {
+          if (!spy.prioritiesTasks[priority]) {
+            this.skip();
+            return;
+          }
+          expect(spy.prioritiesTasks[priority].callCount).toBe(1);
+        });
       });
-      it(`should call ${PREPARING_EACH_ENTITY_PRIORITY} tasks twice`, function () {
-        expect(spy.prioritiesTasks[PREPARING_EACH_ENTITY_PRIORITY].callCount).toBe(2);
+      [CONFIGURING_EACH_ENTITY, PREPARING_EACH_ENTITY, POST_PREPARING_EACH_ENTITY].forEach(priority => {
+        it(`should call ${priority} tasks twice`, function () {
+          if (!spy.prioritiesTasks[priority]) {
+            this.skip();
+            return;
+          }
+          expect(spy.prioritiesTasks[priority].callCount).toBe(2);
+        });
       });
-      it(`should call ${PREPARING_EACH_ENTITY_FIELD_PRIORITY} tasks 3 times`, function () {
-        expect(spy.prioritiesTasks[PREPARING_EACH_ENTITY_FIELD_PRIORITY].callCount).toBe(3);
+      it(`should call ${PREPARING_EACH_ENTITY_FIELD} tasks 3 times`, function () {
+        if (!spy.prioritiesTasks[PREPARING_EACH_ENTITY_FIELD]) {
+          this.skip();
+          return;
+        }
+        expect(spy.prioritiesTasks[PREPARING_EACH_ENTITY_FIELD].callCount).toBe(3);
       });
-      it(`should call ${PREPARING_EACH_ENTITY_RELATIONSHIP_PRIORITY} tasks 3 times`, function () {
-        expect(spy.prioritiesTasks[PREPARING_EACH_ENTITY_RELATIONSHIP_PRIORITY].callCount).toBe(3);
+      it(`should call ${PREPARING_EACH_ENTITY_RELATIONSHIP} tasks 3 times`, function () {
+        if (!spy.prioritiesTasks[PREPARING_EACH_ENTITY_RELATIONSHIP]) {
+          this.skip();
+          return;
+        }
+        expect(spy.prioritiesTasks[PREPARING_EACH_ENTITY_RELATIONSHIP].callCount).toBe(3);
       });
     }
   });
